requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-socketio[asyncio_client]==5.10.0
requests==2.31.0
